refactor(reference): move FASTA index construction into a prepared_index module
Index construction moves out of MultiFastaProvider into a new prepared_index
module, including the protein-FASTA indexing logic factored out into
add_protein_fasta. This is a pure refactor with no behavior change; it exists
so that a future on-disk index artifact deserializes into a type that already
has exactly one other producer, rather than becoming a second one that can
silently diverge.

MultiFastaProvider and from_manifest_inner are not yet routed through
PreparedIndex (that lands in the next commit), so nothing outside its own
tests constructs one yet. PreparedIndex and its impl carry a temporary,
narrowly-scoped #[allow(dead_code)] for that window, with a comment naming
the next commit as the point where it comes off.
